simpleApps/requirements.txt
streamlit>=1.28.0
psycopg[binary,pool]>=3.1.0
databricks-sdk>=0.18.0 
databricks-connect==15.4.*